test: Break apart rspec and minitest
