+++ rails/blueprints/test/rspec/schema_spec
