+++ rails/test/template.rb
@@ -1,9 +1,1 @@
-copy_file 'test/test_helper.rb', force: true
-copy_file 'test/support/circleci.rb'
-copy_file 'test/support/cleaner.rb'
-copy_file 'test/support/factorybot.rb'
-copy_file 'test/support/faker.rb'
-copy_file 'test/support/mailer.rb'
-copy_file 'test/support/rails.rb'
-copy_file 'test/support/retry.rb'
-copy_file 'test/support/simplecov.rb'
+
